rather add functionality to formattedLabel class than to extend it
